+++ src/routes/prefs/+page.svelte
@@ -146,6 +146,14 @@
 			>.
 		</TogglePref>
 		<TogglePref
+			label={m.backendPref_other_server_title()}
+			value={!prefs.backendSkipConfirm}
+			set={set((value, prefs) => (prefs.backendSkipConfirm = !value))}
+		>
+			{m.backendPref_other_server_content()}
+		</TogglePref>
+
+		<TogglePref
 			label={m.prefs_miscellaneous_pullBeforeLaunch_title()}
 			value={prefs.pullBeforeLaunch}
 			set={set((value, prefs) => (prefs.pullBeforeLaunch = value))}
@@ -206,13 +214,6 @@
 
 			{#if games.activeBackends.length > 1}
 				<BackendPref value={gamePrefs.backend} set={set((value) => (gamePrefs!.backend = value))} />
-				<TogglePref
-					label={m.backendPref_other_server_title()}
-					value={!prefs.backendSkipConfirm}
-					set={set((value, prefs) => (prefs.backendSkipConfirm = !value))}
-				>
-					{m.backendPref_other_server_content()}
-				</TogglePref>
 			{/if}
 
 			<ApiKeyPref backend={Backend.Hexium} />