+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+### Added
+
+- Button to view changelogs in the update mods dialog
+
+### Changed
+
+- Cross-source updates are now clearly marked in the update mods dialog
+- Game icons are now sourced from Thunderstore rather than being bundled by Gale
+
 ### Fixed
 
 - Mods sometimes not being installed until Gale is restarted
@@ -9,6 +18,7 @@
 - Clipping issues with the unknown mods banner
 - Mod card lists such as the dependencies and hidden mods dialogs not being sorted correctly
 - Icons and links in update dialog always pointing to Thunderstore
+- Install button being stuck in a disabled state after installing mods
 
 ## 1.19.2 (2026-07-31)